[ThorpS3Client] Extract QuoteStripper and S3ClientObjectLister
