+++ src/main/scala/net/kemitix/s3thorp/awssdk/QuoteStripper.scala
@@ -1,0 +1,7 @@
+package net.kemitix.s3thorp.awssdk
+
+trait QuoteStripper {
+
+  def stripQuotes: Char => Boolean = _ != '"'
+
+}
